+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Added
 
-- Nuvlabox - suspended state
+- Nuvlabox - suspended state and unsuspend operation
 - Module - follow-customer-trial attribute added. This attribute will allow
   vendor to request module trial period to be same as nuvla trial period